@@ -1,82 +1,132 @@
 {
-  "name": "weferral-dashboard",
+  "name": "referal-it",
   "version": "0.1.0",
-  "description": "Weferral provides all the tools needed to start, manage, and grow your affiliate program",
-  "main": "index.js",
+  "author": "Maisnam Raju",
+  "description": "Referral management system",
+  "private": true,
+  "packageManager": "yarn@4.0.0-rc.6",
+  "type": "module",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "npm run start:dev",
-    "build:dev": "node ./build/cli-tools.js --clear dist --create dist && webpack --config ./build/webpack.config.client.dev.js",
-    "build:prod": "node ./build/cli-tools.js --clear dist --create dist && webpack --config ./build/webpack.config.client.prod.js",
-    "start:dev": "node ./build/cli-tools.js --clear dist --create dist && webpack-dev-server --config ./build/webpack.config.client.dev.js",
-    "start:prod": "node ./build/cli-tools.js --clear dist --create dist && webpack-dev-server --config ./build/webpack.config.client.prod.js"
+    "postinstall": "husky install && yarn node ./scripts/postinstall.js",
+    "start": "webpack serve --mode=development",
+    "update-schema": "zx ./scripts/update-schema.js",
+    "lint": "eslint --cache --report-unused-disable-directives .",
+    "test": "jest",
+    "build": "webpack --mode=production",
+    "cf": "zx ./scripts/cf.js"
   },
-  "author": "Ezeh Uche",
-  "license": "ISC",
   "devDependencies": {
     "@babel/cli": "^7.4.4",
     "@babel/core": "^7.4.5",
     "@babel/plugin-proposal-class-properties": "^7.4.4",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
     "@babel/preset-env": "^7.4.5",
-    "@babel/preset-react": "^7.0.0",
-    "autoprefixer": "^9.6.0",
+    "@babel/preset-react": "^7.17.12",
+    "@babel/preset-typescript": "^7.17.12",
+    "@babel/register": "^7.17.7",
+    "@cloudflare/workers-types": "^3.11.0",
+    "@emotion/babel-plugin": "^11.9.2",
+    "@emotion/eslint-plugin": "^11.7.0",
+    "@jest/types": "^28.1.1",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.7",
+    "@types/babel__core": "^7.1.19",
+    "@types/babel__plugin-transform-runtime": "^7.9.2",
+    "@types/babel__preset-env": "^7.9.2",
+    "@types/eslint": "^8.4.2",
+    "@types/js-cookie": "^3.0.2",
+    "@types/prettier": "^2.6.3",
+    "@types/react": "^18.0.12",
+    "@types/react-dom": "^18.0.5",
+    "@types/webpack-bundle-analyzer": "^4.4.1",
+    "@types/webpack-env": "^1.17.0",
+    "@types/webpack-manifest-plugin": "^3.0.5",
+    "@typescript-eslint/eslint-plugin": "^5.27.1",
+    "@typescript-eslint/parser": "^5.27.1",
+    "autoprefixer": "^10.4.7",
     "babel-eslint": "^10.0.1",
+    "babel-jest": "^28.1.1",
     "babel-loader": "^8.0.6",
+    "babel-plugin-import": "^1.13.5",
     "babel-plugin-universal-import": "^4.0.0",
     "circular-dependency-plugin": "^5.0.2",
-    "commander": "^2.20.0",
-    "css-loader": "^3.1.0",
-    "eslint": "~8.16.0",
-    "eslint-config-airbnb": "~19.0.4",
+    "commander": "^9.3.0",
+    "cross-spawn": "^7.0.3",
+    "css-loader": "^6.7.1",
+    "del-cli": "^4.0.1",
+    "envars": "^0.4.0",
+    "eslint": "^8.17.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^8.5.0",
     "eslint-import-resolver-alias": "^1.1.2",
     "eslint-plugin-import": "^2.17.3",
     "eslint-plugin-jsx-a11y": "^6.2.1",
     "eslint-plugin-react": "^7.13.0",
+    "eslint-plugin-react-hooks": "^4.5.0",
+    "execa": "^6.1.0",
     "extract-css-chunks-webpack-plugin": "^4.5.2",
     "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^3.2.0",
+    "html-webpack-plugin": "^5.5.0",
+    "http-proxy-middleware": "^2.0.6",
+    "husky": "^8.0.1",
+    "ignore-webpack-plugin": "^0.3.0",
+    "inline-chunk-html-plugin": "^1.1.1",
+    "jest": "^28.1.1",
     "mkdirp": "^1.0.4",
-    "node-sass": "^4.12.0",
-    "node-sass": "^6.0.1",
-    "optimize-css-assets-webpack-plugin": "^5.0.1",
-    "postcss-loader": "^3.0.0",
+    "node-sass": "^7.0.1",
+    "optimize-css-assets-webpack-plugin": "^6.0.1",
+    "postcss-loader": "^7.0.0",
+    "prettier": "^2.6.2",
     "raw-loader": "^4.0.2",
+    "react-refresh": "^0.13.0",
+    "relay-config": "^12.0.1",
     "rimraf": "^3.0.2",
-    "sass-loader": "^7.1.0",
-    "style-loader": "^0.23.1",
-    "terser-webpack-plugin": "^1.3.0",
-    "webpack": "^4.33.0",
+    "sass-loader": "^13.0.0",
+    "style-loader": "^3.3.1",
+    "terser-webpack-plugin": "^5.3.3",
+    "type-fest": "^2.13.0",
+    "typescript": "^4.7.3",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.73.0",
+    "webpack-bundle-analyzer": "^4.5.0",
     "webpack-cli": "^4.7.2",
-    "webpack-dev-server": "^3.7.1"
+    "webpack-dev-server": "^4.9.2",
+    "webpack-manifest-plugin": "^5.0.0",
+    "wrangler": "^2.0.8",
+    "zx": "^6.2.3"
   },
   "dependencies": {
     "@babel/polyfill": "^7.4.4",
-    "@owczar/dashboard-style--airframe": "^0.1.13",
-    "ag-grid-community": "^21.0.1",
-    "ag-grid-react": "^21.0.1",
-    "axios": "^0.21.0",
+    "@babel/runtime": "^7.18.3",
+    "@cloudflare/kv-asset-handler": "^0.2.0",
+    "@faker-js/faker": "^7.3.0",
+    "@miltonbo/dashboard-style-airframe": "~0.2.0",
+    "ag-grid-community": "^27.3.0",
+    "ag-grid-react": "^27.3.0",
+    "axios": "^0.27.2",
     "bootstrap": "^4.3.1",
     "classnames": "^2.2.6",
-    "csv": "^5.3.2",
+    "copy-webpack-plugin": "^11.0.0",
+    "csv": "^6.1.0",
     "font-awesome": "^4.7.0",
+    "history": "^5.3.0",
     "holderjs": "^2.9.6",
     "immutability-helper": "^3.1.1",
     "isomorphic-fetch": "^3.0.0",
-    "js-cookie": "^2.2.1",
+    "js-cookie": "^3.0.1",
     "js-file-download": "^0.4.12",
     "lodash": "^4.17.11",
     "moment": "^2.29.1",
-    "node-fetch": "^2.6.0",
+    "node-fetch": "^3.2.6",
     "numeral": "^2.0.6",
+    "path-to-regexp": "^6.2.1",
     "prop-types": "^15.7.2",
-    "query-string": "^6.7.0",
-    "rc-slider": "^9.7.2",
-    "react": "^16.8.6",
-    "react-beautiful-dnd": "^11.0.4",
-    "react-big-calendar": "^0.22.0",
+    "query-string": "^7.1.1",
+    "rc-slider": "^10.0.0",
+    "react": "^18.1.0",
+    "react-beautiful-dnd": "^13.1.0",
+    "react-big-calendar": "^0.40.1",
     "react-bootstrap-table": "^4.3.1",
-    "react-bootstrap-table-next": "^3.1.4",
+    "react-bootstrap-table-next": "^4.0.3",
     "react-bootstrap-table2-editor": "^1.2.4",
     "react-bootstrap-table2-filter": "^1.1.9",
     "react-bootstrap-table2-paginator": "^2.0.6",
