@@ -3,16 +3,13 @@ import PropTypes from 'prop-types';
 
 import {
     InputGroup,
-    InputGroupAddon,
     Input
 } from './../../../../components';
 
 // eslint-disable-next-line react/display-name
 const AddonInputFR = React.forwardRef((props, ref) => (
     <InputGroup className={props.className}>
-        <InputGroupAddon addonType="prepend">
-            <i className="fa fa-calendar-o"></i>
-        </InputGroupAddon>
+        <i className="fa fa-calendar-o"></i>
         <Input
             onClick={ props.onClick }
             onChange={ props.onChange }
