@@ -15,7 +15,7 @@ import {
     Media,
     Badge,
     Button,
-    CustomInput,
+    Input,
     StarRating,
     ButtonGroup
 } from '../../../../app/components';
@@ -316,10 +316,10 @@ export class CampaignField extends React.Component {
                 onSelect: this.handleSelect.bind(this),
                 onSelectAll: this.handleSelectAll.bind(this),
                 selectionRenderer: ({ mode, checked, disabled }) => (
-                    <CustomInput type={mode} checked={checked} disabled={disabled} />
+                    <Input type={mode} checked={checked} disabled={disabled} />
                 ),
                 selectionHeaderRenderer: ({ mode, checked, indeterminate }) => (
-                    <CustomInput type={mode} checked={checked} innerRef={el => el && (el.indeterminate = indeterminate)} />
+                    <Input type={mode} checked={checked} innerRef={el => el && (el.indeterminate = indeterminate)} />
                 )
             };
 
