@@ -9,7 +9,7 @@ import {
     EmptyLayout,
     Badge,
     Button,
-    CustomInput,
+    Input,
     ButtonGroup,
     UncontrolledButtonDropdown,
     DropdownToggle,
@@ -208,10 +208,10 @@ export class ParticipantCommission extends React.Component {
                 onSelect: this.handleSelect.bind(this),
                 onSelectAll: this.handleSelectAll.bind(this),
                 selectionRenderer: ({ mode, checked, disabled }) => (
-                    <CustomInput type={mode} checked={checked} disabled={disabled} />
+                    <Input type={mode} checked={checked} disabled={disabled} />
                 ),
                 selectionHeaderRenderer: ({ mode, checked, indeterminate }) => (
-                    <CustomInput type={mode} checked={checked} innerRef={el => el && (el.indeterminate = indeterminate)} />
+                    <Input type={mode} checked={checked} innerRef={el => el && (el.indeterminate = indeterminate)} />
                 )
             };
 
