@@ -2,7 +2,7 @@ import React from 'react';
 import _ from 'lodash';
 import BootstrapTable from 'react-bootstrap-table-next';
 import ToolkitProvider from 'react-bootstrap-table2-toolkit';
-import faker from 'faker/locale/en_US';
+import { faker } from '@faker-js/faker';
 
 import { CustomSearch } from './CustomSearch';
 
