@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import faker from 'faker/locale/en_US';
+import { faker } from '@faker-js/faker';
 import { 
     Media, 
 } from './../../../components';
